+++ R/outcome_fixer.R
@@ -4,6 +4,7 @@
 #' factorised, labelled variable
 #' @param x data frame containing outcome variables
 #' @param who_defined create new factor variable with levels defined by WHO (2013)
+#' @param bin_outcome create new binary factor variable simplifying treatment outcome
 #' @param rm_orig remove original variables - TRUE or FALSE
 #' @param ... further arguments passed to or from other methods
 #' @author Jay Achar
@@ -13,13 +14,15 @@
 
 outcome_fixer <- function(x,
                           who_defined = TRUE,
+                          bin_outcome = TRUE,
                           rm_orig = TRUE,
                           ...) {
 
     # check input
     assert_that(is.data.frame(x),
                 is.flag(rm_orig),
-                is.flag(who_defined))
+                is.flag(who_defined),
+                is.flag(bin_outcome))
 
     # apply correct method
     UseMethod("outcome_fixer", x)